Move `KeyStatusListener` to the `service` package
