+++ android/src/main/kotlin/net/mullvad/mullvadvpn/service/KeyStatusListener.kt
@@ -1,10 +1,9 @@
-package net.mullvad.mullvadvpn.dataproxy
+package net.mullvad.mullvadvpn.service
 
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.GlobalScope
 import kotlinx.coroutines.launch
 import net.mullvad.mullvadvpn.model.KeygenEvent
-import net.mullvad.mullvadvpn.service.MullvadDaemon
 import net.mullvad.talpid.util.EventNotifier
 
 class KeyStatusListener(val daemon: MullvadDaemon) {
